refactor(presence): key cursor/presence color on actor identity
Presence colored peers by the ephemeral peer_id, so a peer's color
changed every session and didn't match their comments or attribution.
Key it on the durable actor identity instead (principal + operator
kind:name) via the shared colorForActorIdentity, falling back to peer_id
only until the actor label arrives.

With presence storing the identity color, the live-presence lookup and
the deterministic fallback in actor-colors converge: cursors, cell dots,
attribution underlines, and comment highlights all read one stable color
per author. identityColorKey moves to the shared editor package; the app
actor-colors helper re-exports it.
