+++ apps/notebook/src/lib/actor-colors.ts
@@ -1,51 +1,24 @@
 /**
- * Single color authority for actor identities, shared across cursors,
- * attribution underlines, and comment highlights so the same author reads as
- * the same color everywhere.
+ * Color authority for actor identities, shared across cursors, attribution
+ * underlines, and comment highlights so the same author reads as the same color
+ * everywhere.
  *
- * Color is keyed on the durable identity (principal + operator kind:name), not
- * the ephemeral per-connection instance id and not the old `peer_id` hash that
- * predates the IdP identity model. That keeps an author's color stable across
- * reconnects and consistent whether they are currently present or not.
- *
- * Stateless by design. A stateful palette allocator would only be needed to
- * guarantee distinct colors among many concurrent actors (hash-collision
- * avoidance); pure hashing is enough for "whose mark is whose."
+ * The deterministic identity hash lives in the shared editor package
+ * (`colorForActorIdentity`), keyed on principal + operator kind:name rather than
+ * the ephemeral instance id or the old peer_id. Presence now stores that same
+ * identity color, so the live-presence lookup and the deterministic fallback
+ * converge; the lookup just lets a connected author resolve before any cursor
+ * has been seen.
  */
 
-import { peerColor } from "@/components/editor/remote-cursors";
+import { colorForActorIdentity, identityColorKey } from "@/components/editor/remote-cursors";
 import { findPeerColorByActorLabel } from "./cursor-registry";
 
-/**
- * Reduce an actor label to its durable identity key.
- *
- * `local:kylekelley/agent:nteract-mcp:6483cc…` → `local:kylekelley/agent:nteract-mcp`
- * `local:kylekelley/desktop:b2c5d701`          → `local:kylekelley/desktop`
- *
- * The trailing segment of an operator is a per-connection/per-device instance
- * id; dropping it groups every session of the same operator under one color.
- */
-export function identityColorKey(actorLabel: string): string {
-  const slash = actorLabel.indexOf("/");
-  if (slash === -1) return actorLabel;
-  const principal = actorLabel.slice(0, slash);
-  const operator = actorLabel.slice(slash + 1);
-  const segments = operator.split(":");
-  const kind = segments[0];
-  const operatorKey =
-    kind === "agent" || kind === "runtime" || kind === "system"
-      ? segments.slice(0, 2).join(":") // kind:name, drop instance id
-      : kind; // e.g. desktop:<device> → desktop
-  return `${principal}/${operatorKey}`;
-}
+export { identityColorKey };
 
-/**
- * Color for a single actor. Prefers the connected peer's live color so a
- * present author's marks match their cursor, falling back to the deterministic
- * identity hash so disconnected authors still get a stable, consistent color.
- */
+/** Color for a single actor: live presence color, else the identity hash. */
 export function colorForActorLabel(actorLabel: string): string {
-  return findPeerColorByActorLabel(actorLabel) ?? peerColor(identityColorKey(actorLabel));
+  return findPeerColorByActorLabel(actorLabel) ?? colorForActorIdentity(actorLabel);
 }
 
 /**
@@ -57,5 +30,5 @@ export function colorForActors(actors: string[]): string {
     const live = findPeerColorByActorLabel(actor);
     if (live) return live;
   }
-  return peerColor(identityColorKey(actors[0] ?? ""));
+  return colorForActorIdentity(actors[0] ?? "");
 }
